Re-arranged code a little, a simple test works now
