+++ server/__tests__/index.test.ts
@@ -1,9 +1,27 @@
 import "jest";
+import { Application } from "express";
+import supertest = require("supertest");
 
-import { App } from "../index";
+import { App } from "../App";
+
+let app: Application;
+let request: supertest.SuperTest<supertest.Test>;
+
 
 describe("Server test suite", () => {
-  test("The app starts without errors", () => {
-    new App();
+
+  beforeAll(() => {
+    app = new App().app;
+  });
+
+  test("The app starts without errors", async () => {
+    request = supertest(app);
+    const response = await request.get("/");
+    expect(response.status).toBe(200);
+    expect(response.type).toBe("text/html");
+  });
+
+  afterAll((done) => {
+    app.close(done);
   });
 });
